+++ src/main/java/com/avab/avab/controller/RecreationController.java
@@ -2,10 +2,12 @@ package com.avab.avab.controller;
 
 import java.util.List;
 
+import org.springframework.data.domain.Page;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.avab.avab.apiPayload.BaseResponse;
@@ -16,30 +18,20 @@ import com.avab.avab.domain.enums.Age;
 import com.avab.avab.domain.enums.Gender;
 import com.avab.avab.domain.enums.Keyword;
 import com.avab.avab.domain.enums.Place;
+import com.avab.avab.dto.response.RecreationResponseDTO.DescriptionDTO;
 import com.avab.avab.dto.response.RecreationResponseDTO.PopularRecreationListDTO;
 import com.avab.avab.dto.response.RecreationResponseDTO.RecreationPreviewListDTO;
 import com.avab.avab.security.handler.annotation.AuthUser;
-import com.avab.avab.converter.RecreationConverter;
-import com.avab.avab.domain.Recreation;
-import com.avab.avab.dto.recreation.RecreationResponseDTO.DescriptionDTO;
-import com.avab.avab.dto.recreation.RecreationResponseDTO.PopularRecreationListDTO;
 import com.avab.avab.service.RecreationService;
-import com.avab.avab.validation.annotation.ValidatePage;
 import com.avab.avab.validation.annotation.ExistRecreation;
+import com.avab.avab.validation.annotation.ValidatePage;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import java.util.List;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Page;
-import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequestMapping("/api/recreations")
@@ -75,29 +67,29 @@ public class RecreationController {
     @Parameter(name = "user", hidden = true)
     @GetMapping("/search")
     public BaseResponse<RecreationPreviewListDTO> searchRecreations(
-        @AuthUser User user,
-        @RequestParam(name = "searchKeyword", required = false) String searchKeyword,
-        @RequestParam(name = "keyword", required = false) List<Keyword> keywords,
-        @RequestParam(name = "participants", required = false) Integer participants,
-        @RequestParam(name = "playTime", required = false) Integer playTime,
-        @RequestParam(name = "place", required = false) List<Place> places,
-        @RequestParam(name = "gender", required = false) List<Gender> genders,
-        @RequestParam(name = "age", required = false) List<Age> ages,
-        @RequestParam(name = "page", required = false, defaultValue = "0") @ValidatePage
-        Integer page) {
+            @AuthUser User user,
+            @RequestParam(name = "searchKeyword", required = false) String searchKeyword,
+            @RequestParam(name = "keyword", required = false) List<Keyword> keywords,
+            @RequestParam(name = "participants", required = false) Integer participants,
+            @RequestParam(name = "playTime", required = false) Integer playTime,
+            @RequestParam(name = "place", required = false) List<Place> places,
+            @RequestParam(name = "gender", required = false) List<Gender> genders,
+            @RequestParam(name = "age", required = false) List<Age> ages,
+            @RequestParam(name = "page", required = false, defaultValue = "0") @ValidatePage
+                    Integer page) {
         Page<Recreation> recreationPage =
-            recreationService.searchRecreations(
-                user,
-                searchKeyword,
-                keywords,
-                participants,
-                playTime,
-                places,
-                genders,
-                ages,
-                page);
+                recreationService.searchRecreations(
+                        user,
+                        searchKeyword,
+                        keywords,
+                        participants,
+                        playTime,
+                        places,
+                        genders,
+                        ages,
+                        page);
 
         return BaseResponse.onSuccess(
-            RecreationConverter.toRecreationPreviewListDTO(recreationPage, user));
+                RecreationConverter.toRecreationPreviewListDTO(recreationPage, user));
     }
 }
