+++ src/main/java/com/avab/avab/dto/response/RecreationResponseDTO.java
@@ -2,6 +2,8 @@ package com.avab.avab.dto.response;
 
 import java.util.List;
 
+import com.avab.avab.domain.enums.Age;
+import com.avab.avab.domain.enums.Gender;
 import com.avab.avab.domain.enums.Keyword;
 
 import io.swagger.v3.oas.annotations.media.Schema;
@@ -54,5 +56,32 @@ public class RecreationResponseDTO {
 
         List<RecreationPreviewDTO> recreationList;
         Integer totalPages;
+    }
+
+    @Builder
+    @Getter
+    @NoArgsConstructor
+    @AllArgsConstructor
+    public static class DescriptionDTO {
+
+        Long recreationId;
+        String summary;
+        List<String> hashTagList;
+        List<Age> ageList;
+        List<String> preparationList;
+        List<WayDTO> wayList;
+        List<Gender> genderList;
+        Integer minParticipants;
+        Integer maxParticipants;
+    }
+
+    @Builder
+    @Getter
+    @NoArgsConstructor
+    @AllArgsConstructor
+    public static class WayDTO {
+
+        String contents;
+        String imageUrl;
     }
 }
