+++ src/main/java/com/avab/avab/service/RecreationService.java
@@ -11,8 +11,6 @@ import com.avab.avab.domain.enums.Gender;
 import com.avab.avab.domain.enums.Keyword;
 import com.avab.avab.domain.enums.Place;
 import com.avab.avab.dto.response.RecreationResponseDTO.PopularRecreationListDTO;
-import com.avab.avab.domain.Recreation;
-import com.avab.avab.dto.recreation.RecreationResponseDTO.PopularRecreationListDTO;
 
 public interface RecreationService {
